+++ components/toast/index.js
@@ -16,6 +16,7 @@ const Toast = function({
   duration = 3000,
   position = 'center',
   hasMask = false,
+  isAutoTop = false,
   parentNode = document.body,
 }) {
   let vm = Toast._instance
@@ -32,6 +33,7 @@ const Toast = function({
         duration,
         position,
         hasMask,
+        isAutoTop,
       },
     }).$mount()
   }
@@ -48,6 +50,7 @@ const Toast = function({
   vm.duration = duration
   vm.position = position
   vm.hasMask = hasMask
+  vm.isAutoTop = isAutoTop
   // vm.visible = true
   // vm.fire()
   vm.show()
@@ -137,15 +140,16 @@ Toast.failed = (content = '', maxTextNum, duration = 3000, hasMask = false, pare
  * @param {string} iconImg
  * @returns {Toast}
  */
-Toast.loading = (content = '', maxTextNum, duration = 0, hasMask = true, parentNode = document.body, iconImg) => {
+Toast.loading = (content = '', maxTextNum, duration = 0, hasMask = true, parentNode = document.body, iconImg, isAutoTop) => {
   return Toast({
     icon: iconImg || 'spinner',
-    iconSvg: iconImg?false:true,
-    iconImg: iconImg?true:false,
+    iconSvg: !iconImg,
+    iconImg: !!iconImg,
     maxTextNum,
     content,
     duration,
     hasMask,
+    isAutoTop,
     parentNode,
   })
 }